refactor: MainPage 분리
